+++ calyx-extra/README.md
@@ -1,0 +1,7 @@
+# Calyx Extra
+
+This defines a bunch of common generators that can come in handy for a bunch of designs.
+
+- math
+    - `gen_ln`: Generates an approximation for `ln` function
+    - `gen_exp`: Generates an approximation for `exp` function
