+++ calyx-extra/calyx-extra/__init__.py
@@ -1,0 +1,3 @@
+"""Generators for commonly used Calyx components"""
+
+__version__ = "0.1.0"
