+++ calyx-extra/pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["flit_core >=3.2,<4"]
+build-backend = "flit_core.buildapi"
+
+[project]
+name = "calyx-extra"
+authors = [{name = "The Calyx Authors"}]
+classifiers = ["License :: OSI Approved :: MIT License"]
+dynamic = ["version", "description"]
+readme = "README.md"
+requires-python = ">=3.4"
+urls.Source = "https://github.com/cucapra/calyx"
+
+[tool.flit.module]
+name = "calyx-extra"
